requirements.txt
# Django stuff
Django
django-tables2
django-extensions
django-filter
django-admin-list-filter-dropdown
django-bootstrap3
django-crispy-forms
django-debug-toolbar
django-enumfield
django-meta

# API stuff
requests

# Database stuff
mysqlclient
psycopg2

# Analytical stuff
django-analytical

# Plot stuff
numpy
plotly
pandas

# Test stuff
faker
factory_boy
coverage
pydotplus
beautifulsoup4
lxml
pytest
pytest-mock
pytest-django
pytest-dotenv
pytest-xdist
django-coverage-plugin

# Dev stuff
black

# Deployment stuff
python-dotenv

# Environment/convenience stuff
ipython
jupyter